Fix parallel processing and reading file list looping infinitely
This change `fflush`es `files_from` before `fork`ing.

Forking a process with open file streams requires that the stream be
`fflush`ed before the `fork` (even read only streams).
Otherwise when the child `exit()`s it will clobber the seek position of
the stream in the parent process, often resulting in an infinite loop as
the end of the file is never found.

Technical details at:
https://stackoverflow.com/questions/50110992/why-does-forking-my-process-cause-the-file-to-be-read-infinitely/50112169#50112169